+++ Makefile
@@ -1,5 +1,5 @@
-main.exe: main.o Menu.o Prenotazione.o Veicolo.o
-	gcc main.o Menu.o Prenotazione.o Veicolo.o -o main.exe
+carsharing.exe: main.o Menu.o Prenotazione.o Veicolo.o Utile.o TabellaHash.o
+	gcc main.o Menu.o Prenotazione.o Veicolo.o Utile.o TabellaHash.o -o carsharing.exe
 
 Menu.o: Menu.c Menu.h
 	gcc -c Menu.c
@@ -10,7 +10,13 @@ Prenotazione.o: Prenotazione.c Prenotazione.h
 Veicolo.o: Veicolo.c Veicolo.h
 	gcc -c Veicolo.c
 
-main.o: main.c Menu.o Prenotazione.o Veicolo.o
+TabellaHash.o: TabellaHash.c TabellaHash.h
+	gcc -c TabellaHash.c
+
+Utile.o: Utile.c Utile.h
+	gcc -c Utile.c
+
+main.o: main.c Menu.o Prenotazione.o Veicolo.o Utile.o TabellaHash.o
 	gcc -c main.c
 
 clean:
