+++ Prenotazione.c
@@ -6,11 +6,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <time.h>
+#include <math.h>
 #include "Veicolo.h"
 #include "Prenotazione.h"
-
-#define HASH_TAGLIA 30
+#include "Utile.h"
 
 struct item {
     int ID; //chiave dell'item prenotazione
@@ -23,31 +22,9 @@ struct item {
     struct item *next; 
 };
 
-struct hash{
-    int taglia;
-    struct item **tabella; 
-};
 
-int ottieniTaglia (TabellaHash t){
-    return t->taglia;
-}
 
-char* ottieniData(){ 
-    time_t t = time (NULL); //ottengo i secondi dal 1 gennaio 1970
-    struct tm* data = localtime (&t); //ottengo la data corrente, ma bisogna formattarla
-    char buffer[20];
-    strftime (buffer, sizeof (buffer), "%d/%m/%Y", data); //formatta la data nel buffer
-    char* dataFormattata = malloc (strlen(buffer) + 1);
 
-    if(dataFormattata == NULL){
-        system("cls|clear");
-        perror("ERRORE: Allocazione memoria data formattata fallita.\n");
-        return NULL;
-    }
-
-    strcpy(dataFormattata, buffer);
-    return dataFormattata;
-}
 /*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: NuovaTabellaHash
  * -----------------------
@@ -76,24 +53,7 @@ char* ottieniData(){
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-TabellaHash NuovaTabellaHash (int taglia){
-    TabellaHash t = malloc (sizeof (struct hash)); //Alloco memoria per la tabella hash
-    if (t == NULL){
-        system("cls | clear");
-        perror ("ERRORE: Allocazione della memoria per la tabella hash fallita.\n");
-        return NULL;
-    }
 
-    t->taglia = taglia+HASH_TAGLIA;
-    t->tabella = calloc ((taglia+HASH_TAGLIA), sizeof (struct item *)); 
-
-    if (t->tabella == NULL){
-        free (t);
-        return NULL;
-    }
-
-    return t;
-}
 
 /*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: FunzioneHash
@@ -123,9 +83,7 @@ TabellaHash NuovaTabellaHash (int taglia){
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-int FunzioneHash(int ID, int taglia) {
-    return ((31 * ID) % taglia);
-}
+
 
 /*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: NuovaPrenotazione
@@ -184,8 +142,8 @@ Prenotazione NuovaPrenotazione (int ID, char* NomeUtente, veicolo c, int i, char
         return NULL;
     }
     
-    p->OrarioSceltoInizio = p->v->orari[i].inizio;
-    p->OrarioSceltoFine = p->v->orari[i].fine;
+    p->OrarioSceltoInizio = ottieniOrarioInizio(c, i);
+    p->OrarioSceltoFine = ottieniOrarioFine(c, i);
     p->CostoNoleggioFinale = costoNoleggio (c, i);
     p->next = NULL;
 
@@ -220,22 +178,7 @@ Prenotazione NuovaPrenotazione (int ID, char* NomeUtente, veicolo c, int i, char
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-int InserisciPrenotazione (TabellaHash t, Prenotazione p){
-    int indice;
-    Prenotazione testa, corrente;
-    indice = FunzioneHash (p->ID, t->taglia);
-    corrente = testa = t->tabella[indice];
-    while (corrente){
-        if (corrente->ID == p->ID){
-            return 0;
-        }
-        corrente = corrente->next;
-    }
-    t->tabella[indice] = p;
-    t->tabella[indice]->next = testa;
 
-    return 1;
-}
 
 /*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: LiberaTabellaHash
@@ -265,15 +208,7 @@ int InserisciPrenotazione (TabellaHash t, Prenotazione p){
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-void LiberaTabellaHash (TabellaHash t){
 
-    for (int i=0; i< t->taglia; i++){
-        LiberaLista (t->tabella[i]);
-    }
-
-    free (t->tabella);
-    free (t);
-}
 
 /*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: LiberaLista
@@ -303,7 +238,7 @@ void LiberaTabellaHash (TabellaHash t){
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-static void LiberaLista (Prenotazione p){
+void LiberaLista (Prenotazione p){
     Prenotazione nuovap;
     while (p != NULL){
         nuovap = p->next;
@@ -344,20 +279,7 @@ static void LiberaLista (Prenotazione p){
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-int AggiornaStorico (Prenotazione p, int indiceVeicolo, int indiceOrario){
-    FILE* file;
-    file = fopen ("StoricoPrenotazioni.txt", "a");
 
-    if (file == NULL){
-        system("cls | clear");
-        perror ("ERRORE: Aggiornamento dello storico da file fallito.\n");
-        return 0;
-    }
-
-    fprintf (file, "\n%s-%s-%.2f-%.2f-%d-%s-%s-%d-%d", p->nomeUtente, p->data, p->OrarioSceltoInizio, p->OrarioSceltoFine, p->ID, p->v->modello, p->v->targa, indiceVeicolo, indiceOrario); // effettuata modifica per gli indici utili nell'inserimento della tabella hash
-
-    return chiudiFile(file);
-}
 
 /*---------------------------------------------------------------------------------------------------------------- 
  * Funzione: stampaPrenotazione
@@ -391,7 +313,7 @@ int AggiornaStorico (Prenotazione p, int indiceVeicolo, int indiceOrario){
 
 
 void stampaPrenotazione (Prenotazione p){ 
-    printf("\nID Prenotazione: %d\nData: %s\nModello: %s\nOrario selezionato: %.2f/%.2f\nCosto noleggio: %.2f euro\n", p->ID, p->data, p->v->modello, p->OrarioSceltoInizio, p->OrarioSceltoFine, p->CostoNoleggioFinale);
+    printf("\nID Prenotazione: %d\nData: %s\nModello: %s\nOrario selezionato: %.2f/%.2f\nCosto noleggio: %.2f euro\n", p->ID, p->data, ottieniModelloPrenotazione(p), p->OrarioSceltoInizio, p->OrarioSceltoFine, p->CostoNoleggioFinale);
 }
 
 /*---------------------------------------------------------------------------------------------------------------- 
@@ -422,188 +344,85 @@ void stampaPrenotazione (Prenotazione p){
  * ---------------------------------------------------------------------------------------------------------------- 
  */
 
-Prenotazione TrovaPrenotazione (TabellaHash t, int ID, int taglia){
-    int indice = FunzioneHash (ID, taglia);
-    Prenotazione p = t->tabella[indice];
-    while (p != NULL){
-        if (p->ID == ID){
-            return p;
-        }
-        p = p->next;
+
+
+
+
+
+
+
+int ottieniID (Prenotazione p){
+    if(p!=NULL){
+        return p->ID;
+    }
+    return -1;
+}
+
+struct item *ottieniNext(Prenotazione p){
+    if(p!=NULL){
+        return p->next;
     }
     return NULL;
 }
 
-TabellaHash RiempiTabellaHashDaFile (veicolo *v){
-    FILE* file = fopen ("StoricoPrenotazioni.txt", "r");
-    if (file == NULL){
-        perror ("ERRORE: Lettura dello storico da file fallito.\n");
-        return NULL;
+char* ottieniNomeUtente(Prenotazione p){
+    if(p!=NULL){
+        return p->nomeUtente;
     }
-
-    char buffer [1024];
-    int contatorePrenotazioni=0;
-
-    while (fgets (buffer, sizeof (buffer), file) != NULL){ contatorePrenotazioni++; }
-    rewind (file);
-
-    TabellaHash t = NuovaTabellaHash (contatorePrenotazioni);
-    if (t == NULL){
-        return NULL;
-    }
-
-    if (contatorePrenotazioni == 0){
-        if (!(chiudiFile(file))){
-            return NULL;
-        }
-        return t;
-    }
-
-    char* dataCorrente = ottieniData();
-    if (dataCorrente == NULL){
-        printf("ERRORE: Ottenimento data locale fallito.\n");
-        LiberaTabellaHash(t);
-        return NULL;
-    }
-
-    fgets(buffer, sizeof(buffer), file); // Leggo a vuoto per partire dalla riga contenente la prenotazione (line 1: "\n")
-    while (fgets (buffer, sizeof (buffer), file) != NULL){
-        char* token = strtok (buffer, "-");
-        if (!(controlloToken (token))){
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            chiudiFile(file);
-            return NULL;
-        }
-
-        char* nomeUtente = strdup (token);
-        if (nomeUtente == NULL){
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            chiudiFile(file);
-            perror ("ERRORE: Inserimento del nome utente della prenotazione nella tabella hash fallito.\n");
-            return NULL;
-        }
-
-        token = strtok (NULL, "-");
-        if (!(controlloToken (token))){
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            chiudiFile(file);
-            free(nomeUtente);
-            return NULL;
-        }
-        
-        char* dataPrenotazione = strdup (token);
-        if (dataPrenotazione == NULL){
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            free(nomeUtente);
-            chiudiFile(file);
-            perror ("ERRORE: Inserimento della data della prenotazione nella tabella hash fallito.\n");
-            return NULL;
-        }
-
-        for (int k=0; k < 3; k++){
-            token = strtok (NULL, "-");
-            if (!(controlloToken (token))){
-                free(dataCorrente);
-                LiberaTabellaHash(t);
-                chiudiFile(file);
-                free(nomeUtente);
-                free(dataPrenotazione);
-                return NULL;
-            }
-        }
-
-        int ID = atoi (token);
-
-        for (int k=0; k < 3; k++){
-            token = strtok (NULL, "-");
-            if (!(controlloToken (token))){
-                free(dataCorrente);
-                LiberaTabellaHash(t);
-                chiudiFile(file);
-                free(nomeUtente);
-                free(dataPrenotazione);
-                return NULL;
-            }
-        }
-
-        int IndiceVeicoloScelto = atoi (token);
-
-        token = strtok (NULL, "-");
-        if (!(controlloToken (token))){
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            chiudiFile(file);
-            free(nomeUtente);
-            free(dataPrenotazione);
-            return NULL;
-        }
-
-        int IndiceOrarioScelto = atoi (token);
-
-        Prenotazione prenotazioneFile = NuovaPrenotazione (ID, nomeUtente, v[IndiceVeicoloScelto], IndiceOrarioScelto, dataPrenotazione);
-        if (prenotazioneFile == NULL){
-            printf("ERRORE: Caricamento della prenotazione da file fallito.\n");
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            chiudiFile(file);
-            free(nomeUtente);
-            free(dataPrenotazione);
-            return NULL;
-        }
-        InserisciPrenotazione(t,prenotazioneFile);
-
-        if (strcmp(dataCorrente, dataPrenotazione) == 0){
-            modificaDisponibilita(v[IndiceVeicoloScelto], IndiceOrarioScelto);
-        }
-        free(nomeUtente);
-        free(dataPrenotazione);
-        }
-
-        if (!(chiudiFile(file))){
-            free(dataCorrente);
-            LiberaTabellaHash(t);
-            return NULL;
-        }
-
-        free(dataCorrente);
-
-    return t;    
+    return NULL;
 }
 
-void LimitaOrariDisponibili (veicolo *v){
-    time_t t = time (NULL); //ottengo i secondi dal 1 gennaio 1970
-    struct tm* data = localtime (&t); //ottengo la data corrente, ma bisogna formattarla
-    int oraCorrente = data->tm_hour;
-    int minutoCorrente = data->tm_min;
-
-    for (int i=0; i<10; i++){
-        for (int k=0; k<8; k++){
-            if (v[i]->orari[k].inizio <= oraCorrente){
-                modificaDisponibilita (v[i], k);
-            }
-        }
+char* ottieniDataPrenotazione(Prenotazione p){
+    if(p!=NULL){
+        return p->data;
     }
+    return NULL;
 }
 
-void StampaPrenotazioneTabellaHash (TabellaHash t, char* nomeUtente){
-    int g = ottieniTaglia(t);
-    int prenotazioniEffettuate = 0;
-
-    for (int i=0; i < g; i++){
-        if (t->tabella[i] != NULL){
-            Prenotazione P = t->tabella[i];
-            if (strcmp (P->nomeUtente, nomeUtente) == 0){
-                prenotazioniEffettuate++;
-                stampaPrenotazione (P);
-            }
-        } 
+float ottieniInizioPrenotazione(Prenotazione p){
+    if(p!=NULL){
+        return p->OrarioSceltoInizio;
     }
+    return -1;
+}
 
-    if (prenotazioniEffettuate == 0){
-        printf ("Non risultano presenti prenotazioni a suo nome.\n");
+float ottieniFinePrenotazione(Prenotazione p){
+    if(p!=NULL){
+        return p->OrarioSceltoFine;
     }
+    return -1;
+}
+
+char* ottieniModelloPrenotazione(Prenotazione p){
+    veicolo temp = p->v;
+
+    if(p!=NULL){
+        return ottieniModello(temp);
+    }
+    return NULL;
+}
+
+char* ottieniTargaPrenotazione(Prenotazione p){
+    veicolo temp = p->v;
+
+    if(p!=NULL){
+        return ottieniTarga(temp);
+    }
+    return NULL;
+}
+
+float costoNoleggio (veicolo v, int k){
+
+    int minutiTotali, ore, minuti;
+    float inizio = ottieniOrarioInizio(v , k), fine = ottieniOrarioFine(v , k);
+
+    ore = (int)inizio;
+    minuti = round((inizio-ore)*100);
+    minutiTotali = (ore*60) + minuti; // MINUTI TOTALI SOLO INIZIALI
+
+    ore = (int)fine;
+    minuti = round((fine-ore)*100);
+    minutiTotali = (ore*60)+minuti - minutiTotali; // MINUTI TOTALI DEL TEMPO DI NOLEGGIO
+
+    return minutiTotali * ((ottieniCostoOrario(v))/60) * verificaSconto(v,k);
 }
