+++ Prenotazione.h
@@ -4,33 +4,30 @@
 #include <time.h>
 #include "Veicolo.h"
 
-typedef struct hash* TabellaHash;
 typedef struct item* Prenotazione;
 
-TabellaHash NuovaTabellaHash (int taglia); //crea e alloca memoria per una tabella hash
-
-int FunzioneHash (int ID, int taglia); //funzione hash che restituisce l'indice nella tabella hash collegato alla chiave, che è l'id
+void LiberaLista (Prenotazione p); // Dealloca una lista collegata nella tabella hash
 
 Prenotazione NuovaPrenotazione (int ID, char* NomeUtente, veicolo c, int i, char* dataPrenotazione); //Crea una nuova prenotazione
 
-int InserisciPrenotazione (TabellaHash t, Prenotazione p); //Inserisce una prenotazione nella tabella hash
 
-void LiberaTabellaHash (TabellaHash t); //Dealloca la memoria occupata dalla tabella hash
-
-static void LiberaLista (Prenotazione p); // Dealloca una lista collegata nella tabella hash
-
-int AggiornaStorico (Prenotazione p, int indiceVeicolo, int indiceOrario); //Aggiorna lo storico prenotazioni con l'ultima prenotazione effettuata
 
 void stampaPrenotazione (Prenotazione p); // stampa i dettagli di una prenotazione
 
-Prenotazione TrovaPrenotazione (TabellaHash t, int ID, int taglia); //cerca una prenotazione nella tabella hash in base all'id
+int ottieniID (Prenotazione p);
 
-TabellaHash RiempiTabellaHashDaFile (veicolo *v); //riempie la tabella hash leggendo le prenotazione dal file StoricoPrenotazioni
+struct item *ottieniNext(Prenotazione p);
 
-int ottieniTaglia (TabellaHash t); // ottiene la taglia della tabella hash
+char* ottieniNomeUtente(Prenotazione p);
 
-char* ottieniData(); // restituisce la data formattata
+char* ottieniDataPrenotazione(Prenotazione p);
 
-void LimitaOrariDisponibili (veicolo *v); //rende gli orari prima della data corrente non disponibili
+float ottieniInizioPrenotazione(Prenotazione p);
 
-void StampaPrenotazioneTabellaHash (TabellaHash t, char* nomeUtente); 
+float ottieniFinePrenotazione(Prenotazione p);
+
+char* ottieniModelloPrenotazione(Prenotazione p);
+
+char* ottieniTargaPrenotazione(Prenotazione p);
+
+float costoNoleggio (veicolo v, int k);
