+++ StoricoPrenotazioni.txt
@@ -1,0 +1,3 @@
+
+Cri-26/05/2025-20.15-22.00-4297-Mercedes Class V-WX901GH-9-7
+Cri-26/05/2025-20.15-22.00-6688-Fiat 500-AB123CD-0-7
